+++ Engine/Renderer/OpenGL/Resource/MeshLoader.h
@@ -9,7 +9,6 @@
 #include "ZeloGLPrerequisites.h"
 #include "Renderer/OpenGL/Resource/GLTexture.h"
 #include "Renderer/OpenGL/Resource/GLMesh.h"
-#include "Renderer/OpenGL/Resource/Material.h"
 #include "Core/ECS/Entity.h"
 
 #include <assimp/scene.h>
